Ajax

+++ routes/projects.js
@@ -1,37 +1,33 @@
 const { Project, Task } = require("../db/models");
-const { asyncHandler, csrfProtection, deleteItem } = require("./utils");
+const { asyncHandler, csrfProtection, deleteItem, findCurrentUser } = require("./utils");
 const express = require("express");
 const { requireAuth } = require("../auth");
 const projectsRouter = express.Router();
 
 // const { projectValidators } = require("./validators"); not created yet
 
+projectsRouter.use(requireAuth);
+
 projectsRouter.get(
 	"/",
-	csrfProtection,
 	asyncHandler(async (req, res) => {
 		const project = await Project.build();
-		const projects = await Project.findAll();
+		const projects = await Project.findAll({ where: { projectOwnerId: findCurrentUser(req.session) } });
 
 		res.render("project", {
 			title: "Projects",
 			project,
 			projects,
-			csrfToken: req.csrfToken(),
 		});
 	})
 );
 
 projectsRouter.post(
 	"/",
-	csrfProtection,
 	// projectValidators,
 	asyncHandler(async (req, res) => {
 		const { projectName } = req.body;
-		const {
-			userAuth: { userId },
-		} = req.session;
-		console.log(userId);
+		userId = findCurrentUser(req.session);
 		const project = await Project.create({ name: projectName, progress: 0, projectOwnerId: userId });
 		res.redirect("/projects");
 	})
@@ -44,8 +40,6 @@ projectsRouter.post(
 
 projectsRouter.post(
 	"/delete/:id(\\d+)",
-	csrfProtection,
-	requireAuth,
 	asyncHandler(async (req, res) => {
 		const projectId = parseInt(req.params.id, 10);
 
@@ -56,12 +50,11 @@ projectsRouter.post(
 
 projectsRouter.get(
 	"/:id(\\d+)",
-	csrfProtection,
 	asyncHandler(async (req, res) => {
 		const tasks = await Task.findAll({ where: { projectId: req.params.id } });
 		console.log(tasks);
 
-		res.render("task", { tasks, title: "Tasks", csrfToken: req.csrfToken() });
+		res.render("task", { tasks, title: "Tasks" });
 	})
 );
